Added dashing and rock abilities and inputs

+++ Assets/Scripts/Extensions/MathfExtensions.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public static class MathfExtensions
 {
-    public static float Sign0(this float value)
+    public static int Sign0(this float value)
     {
         return value == 0 ? 0 : (value > 0 ? 1 : -1);
     }
